// api/model/city-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ای پی آی زومیلا
 * راهنمای استفاده از رابط برنامه نویسی نرم افزار زومیلا
 *
 * The version of the OpenAPI document: v1
 * Contact: jzabihi1980@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CityDto
 */
export interface CityDto {
    /**
     * کد شهر
     * @type {number}
     * @memberof CityDto
     */
    'id'?: number;
    /**
     * کد استان
     * @type {number}
     * @memberof CityDto
     */
    'provinceId'?: number;
    /**
     * نام شهر
     * @type {string}
     * @memberof CityDto
     */
    'name'?: string | null;
}